format code and remove some useless code

+++ srt/PacketSendQueue.cpp
@@ -66,7 +66,7 @@ uint32_t PacketSendQueue::timeLantency() {
     } else {
         dur = first - last;
     }
-    if (dur > (0x01 << 31)) {
+    if (dur > ((uint32_t)0x01 << 31)) {
         TraceL << "cycle timeLantency " << dur;
         dur = 0xffffffff - dur;
     }
